requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
httpx
python-dotenv
